requirements.txt
pyam-iamc >= 2.1  # the `pyam` package is released on pypi under this name
nomenclature-iamc >= 0.16
